% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floor_location.R
\name{floor_location}
\alias{floor_location}
\title{Floors board location}
\usage{
floor_location(a_turn)
}
\description{
This function resets players location counter when they pass Go
}
\examples{
\dontrun{
floor_location(a_turn)
}
}
