\name{hawkins}
\alias{hawkins}
\docType{data}
\title{
hawkins
}
\description{
This data gives x and y variables for the data published in Hawkins'
1994 article.  This data was originally simulated from a standard cubic
polynomial with equally spaced x values between -1 and 1. 
}
\format{
  A data frame with 50 simulated observations on the following 2 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x}}{a numeric vector}
  }
}
\source{
Hawkins, D. M. (1994) Fitting monotonic polynomials to
  data. \emph{Computational Statistics} \bold{9}(3):
  233--247. 
}
\examples{
data(hawkins)
}
\keyword{datasets}
