% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_game.R
\name{simulate_game}
\alias{simulate_game}
\title{Simulate a game of Monopoly}
\usage{
simulate_game(players = 4, turn_limit = 1000)
}
\arguments{
\item{players}{number of players, an integer, defaults to 4}

\item{turn_limit}{maxiumum number of simulated turns, an integer, defaults to 1,000}
}
\description{
Simulate a game of Monopoly, iteratively running take_turn()
}
\examples{
\dontrun{
simulate_game(5, 100)
}
}
