% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_a_chance.R
\name{take_a_chance}
\alias{take_a_chance}
\title{Take a chance}
\usage{
take_a_chance(a_sub_turn)
}
\arguments{
\item{a_sub_turn}{a tibble detailing sub turn informtion, generated as part of take_turn()}
}
\description{
Simulate picking a Chance card from the Chance or Community Chest pile
}
\examples{
\dontrun{
take_a_chance(turn_1)
}
}
