% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_turn.R
\name{take_turn}
\alias{take_turn}
\title{Have a player take their turn}
\usage{
take_turn(last_turn)
}
\arguments{
\item{last_turn}{information about the last turn, a tibble generated either by set_up_game() or take_turn()}
}
\description{
This function simulates a players' turns, based on their previous location on the board.
}
\examples{
\dontrun{
take_turn(player = 1, last_turn = set_up_game())
}
}
